Initial Commit - v1 - Convert, CreateFHIRBundle and POSTtoHAPI functionality - ts
Contains created and converted program commands for the following functionalities:
1. Convert ------------
Description : Converts a .cql (clinical quality language) file to a base64 string and returns it in console or a .txt file, depending on whether output path is provided or not
Inputs:
a. File path (mandatory) - the input file path
b. Output path (optional) - the output file path, if it is provided. If the output path is not provided then just paste the output base64 into the console.
2. CreateFHIRBundle ------------
Description : Creates a FHIR bundle (Fast Healthcare Interoperability Resources) as a json file from an input .cql (clinical quality language) file.A sample .cql file is attached. It will parse details like library name and version to put in the FHIR Bundle and convert the .cql file to base64 and add it to the FHIR bundle file.
Inputs:
a. File path (mandatory) - the input file path
b. Output path (mandatory) - the output file path where the FHIR bundle json is required to be stored
c. Description (mandatory) - the description of the library, which will be added to the FHIR bundle json
d. ipUrl - optional - has default value - Takes in URL to be placed in the library cql fhir bundle

3. POSTtoHAPI ------------
Description: Takes a FHIR bundle (Fast Healthcare Interoperability Resources) as a json file from a path and POSTS (performs a POST request) it to a HAPI-FHIR server.
Inputs :
a. File path (mandatory) - the input file path which is a JSON FHIR Bundle
b. URL (mandatory) - the url where the POST request is to be performed