Added CreateAndPOST command
Created the CreateAndPOST command ---------------
Description :
1. Takes input as cql file and converts it into a FHIR Bundle the way it is done in previous functions
2.  Posts that FHIR Bundle in a URL
Inputs -
a. File path (mandatory) - the input file path
b. Output path (mandatory) - the output file path where the converted fhir json will be stored locally.
c. Description (mandatory) - The description of the library, which will be used in the FHIR bundle json file.
d. URL (mandatory) - the url where the POST request is to be performed. It is also the URL which will be stored in the FHIR json.